Use WebAssembly instead of GopherJS
This makes it more future-proof as GopherJS needs to catch up with the
Golang version, e.g. right now it supports Golang 1.17, and current
version is 1.19.

Currently the WebAssembly support for Golang doesn't support exporting,
and the Go code is run as an application. For this reason the `main`
function isn't allowed to terminate, so we use a channel and the
`finish` function to terminate `main` when we no longer use it.