Handle rego files with multiple rules
The opa inspect command returns one line of json for each rule. When
there's more than one they need to be converted to an array in order
to be parsed successfully as JSON.

This does change the output as well, hence the snapshot change in
the tests. (The diff is small if you ignore the indent change.)